Implementing JWT/JWK-based API security (#483)
This PR (finally) introduces a major feature: Authentication in the gRPC and REST API. 

# Authentication Middleware for gRPC Services

On the server side, it uses an interception feature of [go-grpc-middleware](https://github.com/grpc-ecosystem/go-grpc-middleware) to insert a `grpc.UnaryServerInterceptor` into the gRPC server that requires authentication credentials in each gRPC call. In the `service` package, the function `ConfigureAuth` can be used to create an appropriate [`AuthFunc`](https://pkg.go.dev/github.com/grpc-ecosystem/go-grpc-middleware/auth#AuthFunc) that is needed by the gRPC middleware. The `ConfigureAuth` has two functional-style options to configure the validation of the tokens transmitted by the client. Either a ECDSA public key can directly be specified (using `WithPublicKey`) or JWKS can be used to off-load the validation to an external server (using `WithJWKSURL`).

# Providing Authentications for Client

On the client side, a so called `Authorizer` can be used to insert Bearer tokens in each RPC-call using `grpc.WithPerRPCCredentials`. For now, only one `InternalAuthorizer` exists, which authenticates a client against the Clouditor Authentication Services (using the `Login` call using username and password) and uses the returned access token to provide a bearer token. In the future, we probably want an `OAuthAuthorizer` which performs a complete OAuth flow to retrieve an access token and use that in a future RPC call.

# Usage of Asymmetric Keys in Authentication Service

This PR makes the internal Clouditor Authentication Service a little bit more standard compliant. However, we will probably not ever implement a full OAuth 2.0 server:
- Switched from symmetric keys to sign tokens to asymmetric ECDSA keys
- Provide public ECDS using JWKS using the `ListPublicKeys` RPC call, which is also exposed under the `/.well-known/jwks.json` REST path
- Provide the possibility to persist API key, protected by a password using PKCS#8 and PBES2. Options in the server can be used to configure this (`WithApiKeyPassword`, `WithApiKeyPath`, `WithApiKeySaveOnCreate`)

Co-authored-by: Nico Haas <40119051+lebogg@users.noreply.github.com>