Re-starting old stream channels instead of re-creating them (#977)
* Re-starting old stream channels instead of re-creating them

This PR adds the behaviour of the `StreamsOf` when a channel is broken (e.g. through EOF). Previously, the stream channel was removed and completly created anew. However, this also meant that certain messages that were present at the time of breakage were lost. This PR changes this by putting the message back into the send-queue and recycling the send-queue in the new stream channel.

* Optimized logging in stream