Stopping gRPC server when hitting ctrl+c (#1151)
The gRPC server was still waiting for some connections when hitting ctrl+c to shutdown. This was because we tried to gracefully shutdown, which did not work because streams were not closed. Now streams are closed properly, but we also need to use Stop() because somehow some connections are still active.