Re-organizing Azure discovery file structure (#1306)
In the past we had major problems because of the complexity and the
amount of code in the Azure discoverer in a single file. This proposes a
new scheme for each discoverer subgroup that follows the following
schema:
* `_discover.go` contains the `discoverXXX` functions, which list all
resources in the azure native API
* `_handle.go` contains the `handleXXX` functions, which translate the
azure API objects into ontology objects
* `_properties.go` contains all utility functions that translate
specific, more complex properties, into properties used in the ontology
* `_test.go` contains all the tests. We could possibily also split this
up into the three groups above

Common functions, especially the properties should go into the
`azure_properties.go` to avoid duplication.