Added support for extended non-compliance-details
This PR adds the possibility for metrics to supply additional (non)-compliance details to an assessment result. This is mainly done through tracking the various comparisons done in a metric. An example metric that uses this new system is also supplied.

Potentially the way this result is created inside the metric can be generalized for multiple metrics.