Added support for extended non-compliance-details (#1564)
* Added support for extended non-compliance-details

This PR adds the possibility for metrics to supply additional (non)-compliance details to an assessment result. This is mainly done through tracking the various comparisons done in a metric. An example metric that uses this new system is also supplied.

Potentially the way this result is created inside the metric can be generalized for multiple metrics.

* target_value snake case

* Add test func reencode

* Extend test handle evidence to check comparison results

* Fixed regex

---------

Co-authored-by: Angelika Schneider <angelika.schneider@aisec.fraunhofer.de>